Modify error handling in budgets and rename constant file
